+++ CHANGELOG.md
@@ -28,6 +28,8 @@
 
 ### Fixed
 - LayerSelection: Fixed filtering of layers.
+- LayerTree and `tree.type` `auto`: Layers with empty category in datasets are classified in folder "No category".
+- 
 
 ---